+++ pyriemann/embedding.py
@@ -7,12 +7,22 @@ from scipy.sparse import csr_matrix
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.manifold import spectral_embedding
 
-from .utils.kernel import kernel_riemann
+from .utils import deprecated
+from .utils.kernel import kernel
 from .utils.distance import pairwise_distance
 
 
+@deprecated(
+    "Embedding is deprecated and will be removed in 0.4.0; "
+    "please use SpectralEmbedding."
+)
 class Embedding(BaseEstimator):
-    """Embed SPD matrices into an Euclidean space of smaller dimension.
+    pass
+
+
+class SpectralEmbedding(BaseEstimator):
+    """Spectral embedding of SPD matrices into an Euclidean space of smaller
+    dimension.
 
     It uses Laplacian Eigenmaps [1]_ to embed SPD matrices into an Euclidean
     space. The basic hypothesis is that high-dimensional data lives in a
@@ -81,6 +91,9 @@ class Embedding(BaseEstimator):
             Returns the instance itself.
 
         """
+        _check_dimensions(X,
+                          n_components=self.n_components)
+
         affinity_matrix = self._get_affinity_matrix(X, self.eps)
         embd = spectral_embedding(adjacency=affinity_matrix,
                                   n_components=self.n_components,
@@ -105,7 +118,7 @@ class Embedding(BaseEstimator):
 
         Returns
         -------
-        X_new: ndarray, shape (n_matrices, n_components)
+        X_new : ndarray, shape (n_matrices, n_components)
             Coordinates of embedded matrices.
 
         """
@@ -113,17 +126,19 @@ class Embedding(BaseEstimator):
         return self.embedding_
 
 
-class RiemannLLE(BaseEstimator, TransformerMixin):
-    """Riemannian Locally Linear Embedding (LLE).
+class LocallyLinearEmbedding(BaseEstimator, TransformerMixin):
+    """Locally Linear Embedding (LLE) of SPD matrices.
 
-    Riemannian LLE as proposed in [1]_. LLE is a non-linear,
+    As proposed in [1]_, Locally Linear Embedding (LLE) is a non-linear,
     neighborhood-preserving dimensionality reduction algorithm which
-    consists of three main steps:
-    For each point x,
+    consists of three main steps. For each point x,
+
     1.  find its k nearest neighbors KNN(x) and
     2.  calculate the best reconstruction of x based on its KNN.
     3.  Then calculate a low-dimensional embedding for all points based on
         the weights in step 2.
+
+    This implementation using SPD matrices is based on [2]_.
 
     Parameters
     ----------
@@ -131,6 +146,8 @@ class RiemannLLE(BaseEstimator, TransformerMixin):
         Dimensionality of projected space.
     n_neighbors : int, default: 5
         Number of neighbors for reconstruction of each point.
+    metric : {'riemann', 'logeuclid', 'euclid'}, default: 'riemann'
+        Metric used for KNN and Kernel estimation.
     reg : float, default: 1e-3
         Regularization parameter.
 
@@ -149,14 +166,18 @@ class RiemannLLE(BaseEstimator, TransformerMixin):
 
     References
     ----------
-    .. [1] A. Goh and R. Vidal, "Clustering and dimensionality reduction
-        on Riemannian manifolds", 2008 IEEE Conference on Computer Vision
-        and Pattern Recognition, June 2008
+    .. [1] S. Roweis and L. K. Saul, "Nonlinear Dimensionality Reduction by
+       Locally Linear Embedding", in Science, Vol 290, Issue 5500, pp.
+       2323-2326, 2000.
+    .. [2] A. Goh and R. Vidal, "Clustering and dimensionality reduction
+       on Riemannian manifolds", CVPR 2008
     """
 
-    def __init__(self, n_components=2, n_neighbors=5, reg=1e-3):
+    def __init__(self, n_components=2, n_neighbors=5, metric='riemann',
+                 reg=1e-3):
         self.n_components = n_components
         self.n_neighbors = n_neighbors
+        self.metric = metric
         self.reg = reg
 
     def fit(self, X, y=None):
@@ -176,10 +197,17 @@ class RiemannLLE(BaseEstimator, TransformerMixin):
 
         """
         self.data_ = X
-        self.embedding_, self.error_ = riemann_lle(X,
-                                                   self.n_components,
-                                                   self.n_neighbors,
-                                                   self.reg)
+        _check_dimensions(X,
+                          n_components=self.n_components,
+                          n_neighbors=self.n_neighbors)
+
+        embd, err = locally_linear_embedding(X,
+                                             n_components=self.n_components,
+                                             n_neighbors=self.n_neighbors,
+                                             metric=self.metric,
+                                             reg=self.reg)
+
+        self.embedding_, self.error_ = embd, err
         return self
 
     def transform(self, X, y=None):
@@ -195,14 +223,22 @@ class RiemannLLE(BaseEstimator, TransformerMixin):
 
         Returns
         -------
-        X_: ndarray, shape (n_matrices, n_components)
+        X_new : ndarray, shape (n_matrices, n_components)
             Coordinates of embedded matrices.
         """
-        pairwise_distances = pairwise_distance(X, self.data_, metric='riemann')
+        _check_dimensions(self.data_,
+                          X)
+        pairwise_distances = pairwise_distance(X,
+                                               self.data_,
+                                               metric=self.metric)
         ind = np.array([np.argsort(dist)[1:self.n_neighbors + 1]
                         for dist in pairwise_distances])
 
-        weights = barycenter_weights(X, self.data_, ind, reg=self.reg)
+        weights = barycenter_weights(X,
+                                     self.data_,
+                                     ind,
+                                     metric=self.metric,
+                                     reg=self.reg)
 
         X_new = np.empty((X.shape[0], self.n_components))
         for i in range(X.shape[0]):
@@ -221,20 +257,17 @@ class RiemannLLE(BaseEstimator, TransformerMixin):
 
         Returns
         -------
-        X_: ndarray, shape (n_matrices, n_components)
+        X_new : ndarray, shape (n_matrices, n_components)
             Coordinates of embedded matrices.
         """
-        self.data_ = X
-        self.embedding_, self.error_ = riemann_lle(X,
-                                                   self.n_components,
-                                                   self.n_neighbors,
-                                                   self.reg)
+        self.fit(X)
         return self.embedding_
 
 
-def barycenter_weights(X, Y, indices, reg=1e-3):
+def barycenter_weights(X, Y, indices, metric='riemann', reg=1e-3):
     """Compute Riemannian barycenter weights of X from Y along the first axis.
-    We estimate the weights to assign to each point in Y[indices] to recover
+
+    Estimates the weights to assign to each point in Y[indices] to recover
     the point X[i] by geodesic interpolation. The barycenter weights sum to 1.
 
     Parameters
@@ -245,9 +278,11 @@ def barycenter_weights(X, Y, indices, reg=1e-3):
         Set of SPD matrices.
     indices : ndarray, shape (n_matrices, n_neighbors)
         Indices of the points in Y used to compute the barycenter
-    reg : float, default=1e-3
+    metric : {'riemann', 'logeuclid', 'euclid'}, default: 'riemann'
+        Kernel metric.
+    reg : float, default: 1e-3
         amount of regularization to add for the problem to be
-        well-posed in the case of n_neighbors > n_dim
+        well-posed in the case of n_neighbors > n_channels
 
     Returns
     -------
@@ -257,19 +292,18 @@ def barycenter_weights(X, Y, indices, reg=1e-3):
     Notes
     -----
     .. versionadded:: 0.2.8
-    See sklearn.manifold._locally_linear.barycenter_weights for original code.
     """
-
     n_matrices, n_neighbors = indices.shape
-    assert X.shape[0] == n_matrices, "Number of index-sets in indices must " \
-                                     "match number of matrices in X."
+    msg = f"Number of index-sets in indices (is {n_matrices}) must match " \
+          f"number of matrices in X (is {X.shape[0]})."
+    assert X.shape[0] == n_matrices, msg
 
     B = np.empty((n_matrices, n_neighbors), dtype=X.dtype)
     v = np.ones(n_neighbors, dtype=X.dtype)
 
     for i in range(n_matrices):
         X_neighbors = Y[indices[i]]
-        G = kernel_riemann(X_neighbors, Cref=X[i])
+        G = kernel(X_neighbors, Cref=X[i], metric=metric)
         trace = np.trace(G)
         if trace > 0:
             R = reg * trace
@@ -277,23 +311,26 @@ def barycenter_weights(X, Y, indices, reg=1e-3):
             R = reg
         G.flat[:: n_neighbors + 1] += R
         w = solve(G, v, sym_pos=True)
-        B[i, :] = w / np.sum(w)
+        B[i] = w / np.sum(w)
     return B
 
 
-def riemann_lle(X,
-                n_components=2,
-                n_neighbors=5,
-                reg=1e-3):
-    """Riemannian Locally Linear Embedding (LLE).
+def locally_linear_embedding(X,
+                             *,
+                             n_components=2,
+                             n_neighbors=5,
+                             metric='riemann',
+                             reg=1e-3):
+    """Perform a Locally Linear Embedding (LLE) of SPD matrices.
 
-    Riemannian LLE as proposed in [1]_. LLE is a non-linear, neighborhood-
-    preserving dimensionality reduction algorithm which consists of three
-    main steps:
-    For each point x,
-    1.  find its k nearest neighbors KNN(x) and
-    2.  calculate the best reconstruction of x based on its KNN.
-    3.  Then calculate a low-dimensional embedding for all points based on
+    As proposed in [1]_, Locally Linear Embedding (LLE) is a non-linear,
+    neighborhood-preserving dimensionality reduction algorithm which consists
+    of three main steps. For each point xi,
+
+    1.  find its k nearest neighbors KNN(xi),
+    2.  calculate the best reconstruction of xi based on its
+        k-nearest-neighbors (Eq.9 in [1]_),
+    3.  calculate a low-dimensional embedding for all points based on
         the weights in step 2.
 
     Parameters
@@ -304,14 +341,21 @@ def riemann_lle(X,
         Dimensionality of projected space.
     n_neighbors : int, default: 5
         Number of neighbors for reconstruction of each point.
+    metric : {'riemann', 'logeuclid', 'euclid'}, default: 'riemann'
+        Metric used for KNN and Kernel estimation.
     reg : float, default: 1e-3
         Regularization parameter.
+
+    Returns
+    -------
+    embd : ndarray, shape (n_matrices, n_components)
+        Locally linear embedding of matrices in X.
+    error : float
+        Error of the projected embedding.
 
     Notes
     -----
     .. versionadded:: 0.2.8
-    See sklearn.manifold._locally_linear.locally_linear_embedding for original
-    code.
 
     References
     ----------
@@ -319,49 +363,45 @@ def riemann_lle(X,
         on Riemannian manifolds", 2008 IEEE Conference on Computer Vision
         and Pattern Recognition, June 2008
     """
-
     n_matrices, n_channels, n_channels = X.shape
-    pairwise_distances = pairwise_distance(X, metric='riemann')
+    pairwise_distances = pairwise_distance(X, metric=metric)
     neighbors = np.array([np.argsort(dist)[1:n_neighbors + 1]
                           for dist in pairwise_distances])
 
-    B = barycenter_weights(X, X, neighbors, reg=reg)
+    B = barycenter_weights(X, X, neighbors, metric=metric, reg=reg)
 
     indptr = np.arange(0, n_matrices * n_neighbors + 1, n_neighbors)
     W = csr_matrix((B.ravel(), neighbors.ravel(), indptr), shape=(n_matrices,
                                                                   n_matrices))
     # M = (W - I).T * (W - I) = W.T * W - W.T - W + I
+    # calculated in the two following lines
     M = (W.T * W - W.T - W).toarray()
     M.flat[:: M.shape[0] + 1] += 1
-    return null_space(
-        M,
-        n_components,
-        k_skip=1,
-    )
 
-
-def null_space(M, k, k_skip=1):
-    """Find the null space of a matrix M.
-
-    Parameters
-    ----------
-    M : ndarray, shape (n_matrices, n_matrices)
-        Input covariance matrix: should be symmetric positive semi-definite
-    k : int
-        Number of eigenvalues/vectors to return
-    k_skip : int, default=1
-        Number of low eigenvalues to skip.
-
-    Notes
-    -----
-    .. versionadded:: 0.2.8
-    See sklearn.manifold._locally_linear.null_space for original code.
-    """
-
-    if hasattr(M, "toarray"):
-        M = M.toarray()
     eigen_values, eigen_vectors = eigh(
-        M, eigvals=(k_skip, k + k_skip - 1), overwrite_a=True
+        M, eigvals=(1, n_components), overwrite_a=True
     )
     index = np.argsort(np.abs(eigen_values))
-    return eigen_vectors[:, index], np.sum(eigen_values)
+    embd, error = eigen_vectors[:, index], np.sum(eigen_values)
+
+    return embd, error
+
+
+def _check_dimensions(X, Y=None, n_components=None, n_neighbors=None):
+    n_matrices, n_channels, n_channels = X.shape
+
+    if not isinstance(Y, type(None)):
+        msg = f"Dimension of matrices in data to be transformed must match " \
+              f"dimension of data used for fitting. Expected " \
+              f"{(n_channels, n_channels)}, got {Y.shape[1:]}."
+        assert Y.shape[1:] == (n_channels, n_channels), msg
+
+    if not isinstance(n_neighbors, type(None)):
+        msg = f"n_neighbors (is {n_neighbors}) must be smaller than " \
+              f"n_matrices (is {n_matrices})."
+        assert n_matrices > n_neighbors, msg
+
+    if not isinstance(n_components, type(None)):
+        msg = f"n_components (is {n_components}) must be smaller than " \
+              f"n_matrices (is {n_matrices})."
+        assert n_components < n_matrices, msg